fix for very unusual case in cactus_normalisation
git-svn-id: svn+ssh://hgwdev.cse.ucsc.edu/projects/compbio/svnroot/hausslerlab/recon/branches/2009-09-16/src/sonLib@6415 a8237c23-d22b-0410-ac1a-db40f18c2caa

+++ inc/sonLibRandom.h
@@ -28,7 +28,7 @@ int32_t st_randomInt(int32_t min, int32_t max);
 /*
  * Returns a random value between 0.0 (inclusive) and 1.0 (exclusive).
  */
-float st_random();
+double st_random();
 
 /*
  * Returns a random value from a list.
